+++ src/globals.py
@@ -1,2 +1,2 @@
-build=38
+build=39
 type='dev'